Fixed usage of fields and tags in ProcessData
